oops accidentally made a rank a bigint

+++ chado/modules/companalysis/companalysis.sql
@@ -99,6 +99,6 @@ CREATE TABLE analysisfeatureprop (
     analysisfeature_id bigint NOT NULL REFERENCES analysisfeature(analysisfeature_id) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
     type_id bigint NOT NULL REFERENCES cvterm(cvterm_id) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
     value TEXT,
-    rank bigint NOT NULL,
+    rank int NOT NULL,
     CONSTRAINT analysisfeature_id_type_id_rank UNIQUE(analysisfeature_id, type_id, rank)
 );
